server/requirements.txt
fastapi
uvicorn
websockets
numpy
opencv-python
mediapipe
google-generativeai
gTTS
python-dotenv